chore(clerk-js): Organization should not depend on User components
