+++ packages/clerk-js/src/ui/components/OrganizationProfile/OrganizationSettings.tsx
@@ -1,9 +1,9 @@
+import { BlockButton } from '../../common';
 import { useCoreOrganization } from '../../contexts';
 import { Col, descriptors, Flex, Icon, localizationKeys } from '../../customizables';
 import { Header, IconButton, NavbarMenuButtonRow, OrganizationPreview, ProfileSection } from '../../elements';
 import { Times } from '../../icons';
 import { useRouter } from '../../router';
-import { BlockButton } from '../UserProfile/UserProfileBlockButtons';
 
 export const OrganizationSettings = () => {
   return (
